requirements.txt
streamlit
pandas
pdfplumber
pydantic
openai
google-generativeai
anthropic
groq